Remove Effectful.Internal.MTL module
1. Less orphan instances to worry about.

2. No confusing 'MTL' shown if these are fully qualified in GHC error messages.
