+++ effectful-core/effectful-core.cabal
@@ -92,7 +92,6 @@ library
                      Effectful.Fail
                      Effectful.Internal.Effect
                      Effectful.Internal.Env
-                     Effectful.Internal.MTL
                      Effectful.Internal.Monad
                      Effectful.Internal.Unlift
                      Effectful.Internal.Utils
