+++ effectful-core/src/Effectful.hs
@@ -64,7 +64,6 @@ import Control.Monad.IO.Unlift
 
 import Effectful.Internal.Effect
 import Effectful.Internal.Env
-import Effectful.Internal.MTL ()
 import Effectful.Internal.Monad
 
 -- $intro
